refactor(api-service): centralize agent conversation activity access fixes NV-8576
Make AgentConversationService the single feature-facing interface so activity persistence, sequence, and fan-out rules cannot be bypassed.
