+++ apps/api/src/app/agents/conversation-runtime/conversation/agent-conversation.service.ts
@@ -1,63 +1,57 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { PinoLogger, shortId } from '@novu/application-generic';
 import {
+  ActivityView,
   ConversationActivityEntity,
-  ConversationActivityRepository,
-  ConversationActivitySenderTypeEnum,
-  ConversationActivityToolData,
-  ConversationActivityTypeEnum,
   ConversationChannel,
   ConversationEntity,
   ConversationParticipantTypeEnum,
   ConversationRepository,
   ConversationStatusEnum,
-  isDuplicateKeyError,
 } from '@novu/dal';
-import type { TriggerRecipientsPayload } from '@novu/shared';
-import { AgentChatLiveActivityPublisher } from '../../agent-chat/agent-chat-live-activity.publisher';
-import { mintApprovalActionIds } from '../../shared/tool-approval/mint-approval-action-ids';
-import { ConversationEventSequenceService } from './conversation-event-sequence.service';
+import { getConversationTitle } from './agent-conversation.helpers';
+import type {
+  ConversationActivityContext,
+  PersistAgentActivityParams,
+  PersistAgentMessageResult,
+  PersistInboundMessageParams,
+  PersistMcpConnectionRequestParams,
+  PersistMcpConnectionResultParams,
+  PersistToolApprovalDecisionParams,
+  PersistToolApprovalRequestParams,
+  PersistToolResultParams,
+  PersistTriggerSignalParams,
+  PersistWorkflowOriginHydrationParams,
+  ResolveConversationParams,
+  UpdateMetadataParams,
+} from './agent-conversation.types';
+import { ConversationActivityLedger } from './conversation-activity-ledger';
+import type { PersistRunLifecycleParams } from './run-lifecycle-activity';
 
-export const INBOUND_ATTACHMENT_ONLY_PREVIEW = '[Attachment]';
-export const DEFAULT_CONVERSATION_TITLE = 'Untitled conversation';
+export {
+  AGENT_HISTORY_LIMIT,
+  DEFAULT_CONVERSATION_TITLE,
+  getConversationTitle,
+  getInboundActivityPreview,
+  INBOUND_ATTACHMENT_ONLY_PREVIEW,
+} from './agent-conversation.helpers';
 
-/** Default number of recent activities loaded as conversation history for every runtime. */
-export const AGENT_HISTORY_LIMIT = 50;
-
-/** Stable per-origin identifier for the workflow-origin signal — see `persistWorkflowOriginHydration`. */
-function workflowOriginSignalIdentifier(platformMessageId: string): string {
-  return `workflow-dispatch-origin:${platformMessageId}`;
-}
-
-export function getConversationTitle(firstMessageText: string): string {
-  const trimmed = firstMessageText.trim();
-
-  if (trimmed.length === 0) {
-    return DEFAULT_CONVERSATION_TITLE;
-  }
-
-  return trimmed.slice(0, 200);
-}
-
-export function getInboundActivityPreview(
-  content: string | undefined,
-  options: { richContent?: Record<string, unknown>; hasPlatformAttachments?: boolean } = {}
-): string {
-  const trimmed = content?.trim() ?? '';
-
-  if (trimmed.length > 0) {
-    return trimmed;
-  }
-
-  const attachments = options.richContent?.attachments;
-  const hasStoredAttachments = Array.isArray(attachments) && attachments.length > 0;
-
-  if (hasStoredAttachments || options.hasPlatformAttachments) {
-    return INBOUND_ATTACHMENT_ONLY_PREVIEW;
-  }
-
-  return trimmed;
-}
+export type {
+  ConversationActivityContext,
+  MetadataOp,
+  PersistAgentActivityParams,
+  PersistAgentMessageResult,
+  PersistInboundMessageParams,
+  PersistMcpConnectionRequestParams,
+  PersistMcpConnectionResultParams,
+  PersistToolApprovalDecisionParams,
+  PersistToolApprovalRequestParams,
+  PersistToolResultParams,
+  PersistTriggerSignalParams,
+  PersistWorkflowOriginHydrationParams,
+  ResolveConversationParams,
+  UpdateMetadataParams,
+} from './agent-conversation.types';
 
 export interface CreateOrGetConversationParams {
   environmentId: string;
@@ -85,138 +79,17 @@ export interface CreateOrGetConversationParams {
   contextKeys?: string[];
 }
 
-export interface PersistInboundMessageParams {
-  conversationId: string;
-  platform: string;
-  integrationId: string;
-  platformThreadId: string;
-  senderType: ConversationActivitySenderTypeEnum;
-  senderId: string;
-  senderName?: string;
-  content: string;
-  richContent?: Record<string, unknown>;
-  hasPlatformAttachments?: boolean;
-  platformMessageId?: string;
-  /** Caller-supplied activity identifier; defaults to a server-minted act_* id */
-  identifier?: string;
-  /** Pre-allocated conversation event sequence; minted at persist time when absent */
-  sequence?: number;
-  environmentId: string;
-  organizationId: string;
-}
-
-export interface ConversationActivityContext {
-  conversationId: string;
-  channel: ConversationChannel;
-  agentIdentifier: string;
-  environmentId: string;
-  organizationId: string;
-}
-
-export interface PersistAgentMessageResult {
-  activity: ConversationActivityEntity;
-  /** `false` when the identifier already existed — the caller lost the persist race. */
-  created: boolean;
-}
-
-export interface PersistAgentActivityParams extends ConversationActivityContext {
-  platformMessageId?: string;
-  /** Overrides channel.platformThreadId when delivery returns a different thread ID */
-  platformThreadId?: string;
-  /** Caller-supplied activity identifier; defaults to a server-minted act_* id */
-  identifier?: string;
-  agentName?: string;
-  content: string;
-  richContent?: Record<string, unknown>;
-  /** Pre-allocated conversation event sequence; minted at persist time when absent */
-  sequence?: number;
-}
-
-export interface PersistToolApprovalRequestParams extends ConversationActivityContext {
-  approvalId: string;
-  toolCallId: string;
-  toolName: string;
-  input?: Record<string, unknown>;
-  /** Human-readable preview for the display timeline. */
-  preview?: string;
-  /** When omitted, self-hosted `tool-approval:*` ids are minted. */
-  approveActionId?: string;
-  denyActionId?: string;
-}
-
-export type MetadataOp =
-  | { action: 'set'; key: string; value: unknown }
-  | { action: 'delete'; key: string }
-  | { action: 'clear' };
-
-export interface UpdateMetadataParams extends ConversationActivityContext {
-  currentMetadata: Record<string, unknown>;
-  ops: MetadataOp[];
-}
-
-export interface ResolveConversationParams extends ConversationActivityContext {
-  summary?: string;
-}
-
-export interface PersistTriggerSignalParams extends ConversationActivityContext {
-  workflowId: string;
-  to: TriggerRecipientsPayload;
-  transactionId: string;
-}
-
-export interface PersistWorkflowOriginHydrationParams extends ConversationActivityContext {
-  platformMessageId: string;
-  platformThreadId: string;
-  messageContent: string;
-  signalData: Record<string, unknown>;
-}
-
-export interface PersistToolApprovalDecisionParams extends ConversationActivityContext {
-  approvalId: string;
-  approved: boolean;
-  toolName?: string;
-  actorType:
-    | ConversationActivitySenderTypeEnum.SUBSCRIBER
-    | ConversationActivitySenderTypeEnum.PLATFORM_USER
-    | ConversationActivitySenderTypeEnum.SYSTEM;
-  actorId: string;
-}
-
-export interface PersistToolResultParams extends ConversationActivityContext {
-  toolCallId: string;
-  toolName?: string;
-  /** The tool's output as returned by the model runtime (JSON-serializable). */
-  output: unknown;
-  /** Human-readable preview for the display timeline; defaults to a generic line. */
-  preview?: string;
-}
-
-export interface PersistMcpConnectionRequestParams extends ConversationActivityContext {
-  actionId: string;
-  mcpId: string;
-  displayName: string;
-  authorizeUrl: string;
-  authorizeUrlWithAutoApprove?: string;
-}
-
-export interface PersistMcpConnectionResultParams extends ConversationActivityContext {
-  actionId: string;
-  mcpId: string;
-  status: 'connected' | 'failed';
-  message?: string;
-}
-
 @Injectable()
 export class AgentConversationService {
   constructor(
     private readonly conversationRepository: ConversationRepository,
-    private readonly activityRepository: ConversationActivityRepository,
-    private readonly eventSequenceService: ConversationEventSequenceService,
-    private readonly agentChatLiveActivityPublisher: AgentChatLiveActivityPublisher,
+    private readonly ledger: ConversationActivityLedger,
     private readonly logger: PinoLogger
   ) {
     this.logger.setContext(this.constructor.name);
   }
+
+  // --- Thread ---
 
   getPrimaryChannel(conversation: ConversationEntity): ConversationChannel {
     const channel = conversation.channels?.[0];
@@ -302,113 +175,6 @@ export class AgentConversationService {
     return conversation;
   }
 
-  private async ensureParticipant(conversation: ConversationEntity, params: CreateOrGetConversationParams) {
-    const alreadyPresent = conversation.participants.some(
-      (p) => p.id === params.participantId && p.type === params.participantType
-    );
-    if (alreadyPresent) return;
-
-    const platformIdentity = `${params.platform}:${params.platformUserId}`;
-
-    if (params.participantType === ConversationParticipantTypeEnum.SUBSCRIBER) {
-      const platformUserIdx = conversation.participants.findIndex(
-        (p) => p.type === ConversationParticipantTypeEnum.PLATFORM_USER && p.id === platformIdentity
-      );
-
-      if (platformUserIdx !== -1) {
-        conversation.participants[platformUserIdx] = { type: params.participantType, id: params.participantId };
-
-        this.logger.debug(
-          `Upgraded participant ${platformIdentity} → subscriber ${params.participantId} in conversation ${conversation._id}`
-        );
-      } else {
-        conversation.participants.push({ type: params.participantType, id: params.participantId });
-      }
-    } else {
-      conversation.participants.push({ type: params.participantType, id: params.participantId });
-    }
-
-    await this.conversationRepository.updateParticipants(
-      params.environmentId,
-      params.organizationId,
-      conversation._id,
-      conversation.participants
-    );
-  }
-
-  async persistInboundMessage(params: PersistInboundMessageParams): Promise<ConversationActivityEntity> {
-    const content = params.content ?? '';
-    const preview = getInboundActivityPreview(content, {
-      richContent: params.richContent,
-      hasPlatformAttachments: params.hasPlatformAttachments,
-    });
-    const identifier = params.identifier ?? `act_${shortId(12)}`;
-    const sequence = await this.resolveEventSequence(
-      params.conversationId,
-      params.environmentId,
-      params.organizationId,
-      params.sequence
-    );
-
-    try {
-      const [activity] = await Promise.all([
-        this.activityRepository.createUserActivity({
-          identifier,
-          conversationId: params.conversationId,
-          platform: params.platform,
-          integrationId: params.integrationId,
-          platformThreadId: params.platformThreadId,
-          senderType: params.senderType,
-          senderId: params.senderId,
-          senderName: params.senderName,
-          content,
-          richContent: params.richContent,
-          platformMessageId: params.platformMessageId,
-          sequence,
-          environmentId: params.environmentId,
-          organizationId: params.organizationId,
-        }),
-        this.conversationRepository.touchActivity(
-          params.environmentId,
-          params.organizationId,
-          params.conversationId,
-          preview
-        ),
-      ]);
-
-      return activity;
-    } catch (err) {
-      if (params.identifier && isDuplicateKeyError(err)) {
-        const existing = await this.activityRepository.findOne(
-          {
-            _environmentId: params.environmentId,
-            _conversationId: params.conversationId,
-            identifier: params.identifier,
-          },
-          '*'
-        );
-
-        if (existing) {
-          return existing;
-        }
-      }
-
-      throw err;
-    }
-  }
-
-  async findSourceActivity(
-    environmentId: string,
-    conversationId: string,
-    platformMessageId: string
-  ): Promise<ConversationActivityEntity | null> {
-    return this.activityRepository.findByPlatformMessageId(environmentId, conversationId, platformMessageId);
-  }
-
-  async countAgentMessages(environmentId: string, conversationId: string): Promise<number> {
-    return this.activityRepository.countAgentMessages(environmentId, conversationId);
-  }
-
   async getConversation(
     conversationId: string,
     environmentId: string,
@@ -489,205 +255,6 @@ export class AgentConversationService {
     );
   }
 
-  async findAgentMessageByIdentifier(
-    environmentId: string,
-    conversationId: string,
-    identifier: string
-  ): Promise<ConversationActivityEntity | null> {
-    return this.activityRepository.findOne(
-      {
-        _environmentId: environmentId,
-        _conversationId: conversationId,
-        identifier,
-        type: ConversationActivityTypeEnum.MESSAGE,
-      },
-      '*'
-    );
-  }
-
-  async persistAgentMessage(params: PersistAgentActivityParams): Promise<PersistAgentMessageResult> {
-    try {
-      const activity = await this.persistAgentActivity(params, ConversationActivityTypeEnum.MESSAGE, 'activity');
-
-      return { activity, created: true };
-    } catch (err) {
-      if (params.identifier && isDuplicateKeyError(err)) {
-        this.logger.warn(
-          { identifier: params.identifier, conversationId: params.conversationId },
-          'Agent message activity already recorded (duplicate identifier)'
-        );
-
-        const existing = await this.activityRepository.findOne(
-          {
-            _environmentId: params.environmentId,
-            _conversationId: params.conversationId,
-            identifier: params.identifier,
-          },
-          '*'
-        );
-
-        if (existing) {
-          return { activity: existing, created: false };
-        }
-      }
-
-      throw err;
-    }
-  }
-
-  /** Records the platform-native message id after a successful post on a persist-first delivery. */
-  async setAgentMessagePlatformMessageId(params: {
-    environmentId: string;
-    organizationId: string;
-    conversationId: string;
-    activityId: string;
-    platformMessageId: string;
-  }): Promise<void> {
-    await this.activityRepository.update(
-      {
-        _environmentId: params.environmentId,
-        _organizationId: params.organizationId,
-        _conversationId: params.conversationId,
-        _id: params.activityId,
-      },
-      { $set: { platformMessageId: params.platformMessageId } }
-    );
-  }
-
-  /** Compensating delete when the platform post fails, so a retry can re-claim the identifier. */
-  async deleteAgentMessage(params: {
-    environmentId: string;
-    organizationId: string;
-    conversationId: string;
-    activityId: string;
-  }): Promise<void> {
-    await this.activityRepository.findOneAndDelete({
-      _environmentId: params.environmentId,
-      _organizationId: params.organizationId,
-      _conversationId: params.conversationId,
-      _id: params.activityId,
-    });
-  }
-
-  async persistToolApprovalRequest(params: PersistToolApprovalRequestParams): Promise<ConversationActivityEntity> {
-    const toolName = params.toolName;
-    const sequence = await this.resolveEventSequence(
-      params.conversationId,
-      params.environmentId,
-      params.organizationId
-    );
-    const actionIds =
-      params.approveActionId && params.denyActionId
-        ? { approveActionId: params.approveActionId, denyActionId: params.denyActionId }
-        : mintApprovalActionIds({ approvalId: params.approvalId });
-
-    const activity = await this.activityRepository.createToolActivity({
-      identifier: `act_${shortId(12)}`,
-      conversationId: params.conversationId,
-      platform: params.channel.platform,
-      integrationId: params.channel._integrationId,
-      platformThreadId: params.channel.platformThreadId,
-      senderType: ConversationActivitySenderTypeEnum.AGENT,
-      senderId: params.agentIdentifier,
-      content: params.preview ?? `Approval required: ${toolName}`,
-      type: ConversationActivityTypeEnum.TOOL_APPROVAL_REQUEST,
-      toolData: {
-        approvalId: params.approvalId,
-        toolCallId: params.toolCallId,
-        toolName: params.toolName,
-        input: params.input,
-        approveActionId: actionIds.approveActionId,
-        denyActionId: actionIds.denyActionId,
-      },
-      sequence,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-    });
-
-    await this.agentChatLiveActivityPublisher.emitPersistedClientEvent({
-      channel: params.channel,
-      conversationId: params.conversationId,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-      agentIdentifier: params.agentIdentifier,
-      activity,
-    });
-
-    return activity;
-  }
-
-  /** Links a delivered approval card message to its ledger row (for platform edits). */
-  async linkToolApprovalRequestCard(params: {
-    environmentId: string;
-    organizationId: string;
-    conversationId: string;
-    activityId: string;
-    platformMessageId: string;
-  }): Promise<void> {
-    await this.activityRepository.update(
-      {
-        _environmentId: params.environmentId,
-        _organizationId: params.organizationId,
-        _conversationId: params.conversationId,
-        _id: params.activityId,
-        type: ConversationActivityTypeEnum.TOOL_APPROVAL_REQUEST,
-      },
-      { $set: { platformMessageId: params.platformMessageId } }
-    );
-  }
-
-  async persistAgentEdit(params: PersistAgentActivityParams): Promise<ConversationActivityEntity> {
-    return this.persistAgentActivity(params, ConversationActivityTypeEnum.EDIT, 'preview');
-  }
-
-  async persistAgentDelete(params: PersistAgentActivityParams): Promise<ConversationActivityEntity> {
-    return this.persistAgentActivity(params, ConversationActivityTypeEnum.DELETE, 'preview');
-  }
-
-  private async persistAgentActivity(
-    params: PersistAgentActivityParams & {
-      toolData?: ConversationActivityToolData;
-    },
-    type: ConversationActivityTypeEnum,
-    touch: 'activity' | 'preview'
-  ): Promise<ConversationActivityEntity> {
-    const threadId = params.platformThreadId ?? params.channel.platformThreadId;
-    const sequence = await this.resolveEventSequence(
-      params.conversationId,
-      params.environmentId,
-      params.organizationId,
-      params.sequence
-    );
-
-    const touchFn =
-      touch === 'activity'
-        ? this.conversationRepository.touchActivity.bind(this.conversationRepository)
-        : this.conversationRepository.touchPreview.bind(this.conversationRepository);
-
-    const [activity] = await Promise.all([
-      this.activityRepository.createAgentActivity({
-        identifier: params.identifier ?? `act_${shortId(12)}`,
-        conversationId: params.conversationId,
-        platform: params.channel.platform,
-        integrationId: params.channel._integrationId,
-        platformThreadId: threadId,
-        platformMessageId: params.platformMessageId,
-        agentId: params.agentIdentifier,
-        senderName: params.agentName,
-        content: params.content,
-        richContent: params.richContent,
-        toolData: params.toolData,
-        type,
-        sequence,
-        environmentId: params.environmentId,
-        organizationId: params.organizationId,
-      }),
-      touchFn(params.environmentId, params.organizationId, params.conversationId, params.content),
-    ]);
-
-    return activity;
-  }
-
   async updateMetadata(params: UpdateMetadataParams): Promise<void> {
     let merged: Record<string, unknown> = { ...(params.currentMetadata ?? {}) };
     const descriptions: string[] = [];
@@ -721,17 +288,10 @@ export class AgentConversationService {
         params.conversationId,
         merged
       ),
-      this.activityRepository.createSignalActivity({
-        identifier: `act_${shortId(12)}`,
-        conversationId: params.conversationId,
-        platform: params.channel.platform,
-        integrationId: params.channel._integrationId,
-        platformThreadId: params.channel.platformThreadId,
-        agentId: params.agentIdentifier,
+      this.ledger.persistMetadataSignal({
+        ...params,
         content: `Metadata updated: ${descriptions.join(', ')}`,
-        signalData: { type: 'metadata', payload: merged },
-        environmentId: params.environmentId,
-        organizationId: params.organizationId,
+        payload: merged,
       }),
     ]);
   }
@@ -753,268 +313,218 @@ export class AgentConversationService {
         new Date().toISOString()
       ),
       this.conversationRepository.clearExternalSessionId(params.environmentId, params.conversationId),
-      this.activityRepository.createSignalActivity({
-        identifier: `act_${shortId(12)}`,
-        conversationId: params.conversationId,
-        platform: params.channel.platform,
-        integrationId: params.channel._integrationId,
-        platformThreadId: params.channel.platformThreadId,
-        agentId: params.agentIdentifier,
+      this.ledger.persistResolveSignal({
+        ...params,
         content: params.summary ?? 'Conversation resolved',
-        signalData: { type: 'resolve', payload: params.summary ? { summary: params.summary } : undefined },
-        environmentId: params.environmentId,
-        organizationId: params.organizationId,
       }),
     ]);
   }
 
-  /**
-   * Persist a tool-approval decision as a signal activity so it becomes part of
-   * the durable transcript. Self-hosted (stateless) agents reconstruct the resume
-   * message list from history via `toModelMessages`, so the decision must live in
-   * the transcript — not only in the ephemeral approval card.
-   */
-  async persistToolApprovalDecision(params: PersistToolApprovalDecisionParams): Promise<ConversationActivityEntity> {
-    const toolName = params.toolName ?? 'tool call';
-    const sequence = await this.resolveEventSequence(
-      params.conversationId,
-      params.environmentId,
-      params.organizationId
-    );
+  // --- Messages ---
 
-    const activity = await this.activityRepository.createToolActivity({
-      identifier: `act_${shortId(12)}`,
-      conversationId: params.conversationId,
-      platform: params.channel.platform,
-      integrationId: params.channel._integrationId,
-      platformThreadId: params.channel.platformThreadId,
-      senderType: params.actorType,
-      senderId: params.actorId,
-      content: params.approved ? `Approved ${toolName}` : `Denied ${toolName}`,
-      type: ConversationActivityTypeEnum.TOOL_APPROVAL_DECISION,
-      toolData: { approvalId: params.approvalId, approved: params.approved, toolName: params.toolName },
-      sequence,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-    });
-
-    await this.agentChatLiveActivityPublisher.emitPersistedClientEvent({
-      channel: params.channel,
-      conversationId: params.conversationId,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-      agentIdentifier: params.agentIdentifier,
-      activity,
-    });
-
-    return activity;
+  async persistInboundMessage(params: PersistInboundMessageParams): Promise<ConversationActivityEntity> {
+    return this.ledger.persistInboundMessage(params);
   }
 
-  async persistToolResult(params: PersistToolResultParams): Promise<void> {
-    const sequence = await this.resolveEventSequence(
-      params.conversationId,
-      params.environmentId,
-      params.organizationId
-    );
-
-    const activity = await this.activityRepository.createToolActivity({
-      identifier: `act_${shortId(12)}`,
-      conversationId: params.conversationId,
-      platform: params.channel.platform,
-      integrationId: params.channel._integrationId,
-      platformThreadId: params.channel.platformThreadId,
-      senderType: ConversationActivitySenderTypeEnum.AGENT,
-      senderId: params.agentIdentifier,
-      content: params.preview ?? `Tool result: ${params.toolName ?? params.toolCallId}`,
-      type: ConversationActivityTypeEnum.TOOL_RESULT,
-      toolData: { toolCallId: params.toolCallId, toolName: params.toolName, output: params.output },
-      sequence,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-    });
-
-    await this.agentChatLiveActivityPublisher.emitPersistedClientEvent({
-      channel: params.channel,
-      conversationId: params.conversationId,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-      agentIdentifier: params.agentIdentifier,
-      activity,
-    });
+  async persistAgentMessage(params: PersistAgentActivityParams): Promise<PersistAgentMessageResult> {
+    return this.ledger.persistAgentMessage(params);
   }
 
-  async persistMcpConnectionRequest(params: PersistMcpConnectionRequestParams): Promise<ConversationActivityEntity> {
-    const activity = await this.persistAgentActivity(
-      {
-        ...params,
-        identifier: `mcp-connection:${params.actionId}:request`,
-        content: `Connect ${params.displayName}`,
-        richContent: {
-          mcpConnection: {
-            actionId: params.actionId,
-            mcpId: params.mcpId,
-            displayName: params.displayName,
-            authorizeUrl: params.authorizeUrl,
-            authorizeUrlWithAutoApprove: params.authorizeUrlWithAutoApprove,
-          },
-        },
-      },
-      ConversationActivityTypeEnum.MCP_CONNECTION_REQUEST,
-      'activity'
-    );
-
-    await this.agentChatLiveActivityPublisher.emitPersistedClientEvent({
-      channel: params.channel,
-      conversationId: params.conversationId,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-      agentIdentifier: params.agentIdentifier,
-      activity,
-    });
-
-    return activity;
+  async setAgentMessagePlatformMessageId(params: {
+    environmentId: string;
+    organizationId: string;
+    conversationId: string;
+    activityId: string;
+    platformMessageId: string;
+  }): Promise<void> {
+    return this.ledger.setAgentMessagePlatformMessageId(params);
   }
 
-  async persistMcpConnectionResult(params: PersistMcpConnectionResultParams): Promise<ConversationActivityEntity> {
-    const activity = await this.persistAgentActivity(
-      {
-        ...params,
-        identifier: `mcp-connection:${params.actionId}:result`,
-        content: params.status === 'connected' ? 'Connection completed' : (params.message ?? 'Connection failed'),
-        richContent: {
-          mcpConnection: {
-            actionId: params.actionId,
-            mcpId: params.mcpId,
-            status: params.status,
-            message: params.message,
-          },
-        },
-      },
-      ConversationActivityTypeEnum.MCP_CONNECTION_RESULT,
-      'activity'
-    );
-
-    await this.agentChatLiveActivityPublisher.emitPersistedClientEvent({
-      channel: params.channel,
-      conversationId: params.conversationId,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-      agentIdentifier: params.agentIdentifier,
-      activity,
-    });
-
-    return activity;
+  async deleteAgentMessage(params: {
+    environmentId: string;
+    organizationId: string;
+    conversationId: string;
+    activityId: string;
+  }): Promise<void> {
+    return this.ledger.deleteAgentMessage(params);
   }
 
-  /**
-   * Whoever makes the event real first mints its sequence: live delivery paths
-   * (web) mint before emitting and pass the value here; everything else gets
-   * one at persist time. Channel-agnostic — every conversation is sequenced.
-   */
-  private async resolveEventSequence(
-    conversationId: string,
-    environmentId: string,
-    organizationId: string,
-    sequence?: number
-  ): Promise<number> {
-    if (sequence !== undefined) {
-      return sequence;
-    }
-
-    return this.eventSequenceService.mint({
-      environmentId,
-      organizationId,
-      conversationId,
-    });
+  async persistAgentEdit(params: PersistAgentActivityParams): Promise<ConversationActivityEntity> {
+    return this.ledger.persistAgentEdit(params);
   }
 
-  async persistTriggerSignal(params: PersistTriggerSignalParams): Promise<void> {
-    await this.activityRepository.createSignalActivity({
-      identifier: `act_${shortId(12)}`,
-      conversationId: params.conversationId,
-      platform: params.channel.platform,
-      integrationId: params.channel._integrationId,
-      platformThreadId: params.channel.platformThreadId,
-      agentId: params.agentIdentifier,
-      content: `Triggered workflow: ${params.workflowId}`,
-      signalData: {
-        type: 'trigger',
-        payload: {
-          workflowId: params.workflowId,
-          to: params.to,
-          transactionId: params.transactionId,
-        },
-      },
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-    });
+  async persistAgentDelete(params: PersistAgentActivityParams): Promise<ConversationActivityEntity> {
+    return this.ledger.persistAgentDelete(params);
   }
 
-  /**
-   * Whether this origin is already in the transcript. The signal is the final write of
-   * `persistWorkflowOriginHydration`, so a partially applied hydration reads as not hydrated.
-   */
+  async persistWorkflowOriginHydration(params: PersistWorkflowOriginHydrationParams): Promise<void> {
+    return this.ledger.persistWorkflowOriginHydration(params);
+  }
+
   async isWorkflowOriginHydrated(
     environmentId: string,
     conversationId: string,
     platformMessageId: string
   ): Promise<boolean> {
-    const count = await this.activityRepository.count(
-      {
-        _environmentId: environmentId,
-        _conversationId: conversationId,
-        identifier: workflowOriginSignalIdentifier(platformMessageId),
-      },
-      1
-    );
-
-    return count > 0;
+    return this.ledger.isWorkflowOriginHydrated(environmentId, conversationId, platformMessageId);
   }
 
-  /**
-   * Persist the workflow-origin message + signal. Stable `workflow-dispatch-*`
-   * identifiers collide on the unique index under concurrency; both writes are
-   * duplicate-key tolerant so a retry after a partial success is idempotent.
-   */
-  async persistWorkflowOriginHydration(params: PersistWorkflowOriginHydrationParams): Promise<void> {
-    await this.persistAgentMessage({
-      conversationId: params.conversationId,
-      channel: params.channel,
-      agentIdentifier: params.agentIdentifier,
-      environmentId: params.environmentId,
-      organizationId: params.organizationId,
-      platformMessageId: params.platformMessageId,
-      platformThreadId: params.platformThreadId,
-      identifier: `workflow-dispatch-msg:${params.platformMessageId}`,
-      content: params.messageContent,
-    });
+  async listForView(params: {
+    view: ActivityView;
+    environmentId: string;
+    organizationId: string;
+    conversationId: string;
+    limit?: number;
+    before?: string;
+  }): Promise<{ data: ConversationActivityEntity[]; hasMore: boolean }> {
+    return this.ledger.listForView(params);
+  }
 
-    try {
-      await this.activityRepository.createSignalActivity({
-        identifier: workflowOriginSignalIdentifier(params.platformMessageId),
-        conversationId: params.conversationId,
-        platform: params.channel.platform,
-        integrationId: params.channel._integrationId,
-        platformThreadId: params.platformThreadId,
-        agentId: params.agentIdentifier,
-        content: `Workflow origin: ${String(params.signalData.workflowIdentifier ?? 'unknown')}`,
-        signalData: {
-          type: 'workflow_origin',
-          payload: params.signalData,
-        },
-        platformMessageId: params.platformMessageId,
-        environmentId: params.environmentId,
-        organizationId: params.organizationId,
-      });
-    } catch (err) {
-      if (!isDuplicateKeyError(err)) {
-        throw err;
-      }
+  // --- Tools ---
 
-      this.logger.warn(
-        { platformMessageId: params.platformMessageId, conversationId: params.conversationId },
-        'Workflow origin already hydrated'
+  async persistToolApprovalRequest(params: PersistToolApprovalRequestParams): Promise<ConversationActivityEntity> {
+    return this.ledger.persistToolApprovalRequest(params);
+  }
+
+  async linkToolApprovalRequestCard(params: {
+    environmentId: string;
+    organizationId: string;
+    conversationId: string;
+    activityId: string;
+    platformMessageId: string;
+  }): Promise<void> {
+    return this.ledger.linkToolApprovalRequestCard(params);
+  }
+
+  async persistToolApprovalDecision(params: PersistToolApprovalDecisionParams): Promise<ConversationActivityEntity> {
+    return this.ledger.persistToolApprovalDecision(params);
+  }
+
+  async persistToolResult(params: PersistToolResultParams): Promise<void> {
+    return this.ledger.persistToolResult(params);
+  }
+
+  async persistMcpConnectionRequest(params: PersistMcpConnectionRequestParams): Promise<ConversationActivityEntity> {
+    return this.ledger.persistMcpConnectionRequest(params);
+  }
+
+  async persistMcpConnectionResult(params: PersistMcpConnectionResultParams): Promise<ConversationActivityEntity> {
+    return this.ledger.persistMcpConnectionResult(params);
+  }
+
+  async persistTriggerSignal(params: PersistTriggerSignalParams): Promise<void> {
+    return this.ledger.persistTriggerSignal(params);
+  }
+
+  async persistRunLifecycle(params: PersistRunLifecycleParams): Promise<ConversationActivityEntity | null> {
+    return this.ledger.persistRunLifecycle(params);
+  }
+
+  // --- Lookups ---
+
+  async findByPlatformMessageId(
+    environmentId: string,
+    conversationId: string,
+    platformMessageId: string
+  ): Promise<ConversationActivityEntity | null> {
+    return this.ledger.findByPlatformMessageId(environmentId, conversationId, platformMessageId);
+  }
+
+  async findSourceActivity(
+    environmentId: string,
+    conversationId: string,
+    platformMessageId: string
+  ): Promise<ConversationActivityEntity | null> {
+    return this.findByPlatformMessageId(environmentId, conversationId, platformMessageId);
+  }
+
+  async countAgentMessages(environmentId: string, conversationId: string): Promise<number> {
+    return this.ledger.countAgentMessages(environmentId, conversationId);
+  }
+
+  async findAgentMessageByIdentifier(
+    environmentId: string,
+    conversationId: string,
+    identifier: string
+  ): Promise<ConversationActivityEntity | null> {
+    return this.ledger.findAgentMessageByIdentifier(environmentId, conversationId, identifier);
+  }
+
+  async findToolActivitiesByPlanMessageId(
+    environmentId: string,
+    conversationId: string,
+    planMessageId: string
+  ): Promise<ConversationActivityEntity[]> {
+    return this.ledger.findToolActivitiesByPlanMessageId(environmentId, conversationId, planMessageId);
+  }
+
+  async persistToolUseSignal(
+    params: ConversationActivityContext & { content: string; payload: Record<string, unknown> }
+  ): Promise<void> {
+    return this.ledger.persistToolUseSignal(params);
+  }
+
+  async enrichToolUseSignal(params: {
+    environmentId: string;
+    organizationId: string;
+    conversationId: string;
+    activityId: string;
+    content: string;
+    payload: Record<string, unknown>;
+  }): Promise<void> {
+    return this.ledger.enrichToolUseSignal(params);
+  }
+
+  async repointSubscriberSender(params: {
+    environmentId: string;
+    organizationId: string;
+    fromSubscriberId: string;
+    toSubscriberId: string;
+  }): Promise<number> {
+    return this.ledger.repointSubscriberSender(params);
+  }
+
+  // --- Sequence ---
+
+  async mintEventSequence(params: {
+    environmentId: string;
+    organizationId: string;
+    conversationId: string;
+  }): Promise<number> {
+    return this.ledger.mint(params);
+  }
+
+  private async ensureParticipant(conversation: ConversationEntity, params: CreateOrGetConversationParams) {
+    const alreadyPresent = conversation.participants.some(
+      (p) => p.id === params.participantId && p.type === params.participantType
+    );
+    if (alreadyPresent) return;
+
+    const platformIdentity = `${params.platform}:${params.platformUserId}`;
+
+    if (params.participantType === ConversationParticipantTypeEnum.SUBSCRIBER) {
+      const platformUserIdx = conversation.participants.findIndex(
+        (p) => p.type === ConversationParticipantTypeEnum.PLATFORM_USER && p.id === platformIdentity
       );
+
+      if (platformUserIdx !== -1) {
+        conversation.participants[platformUserIdx] = { type: params.participantType, id: params.participantId };
+
+        this.logger.debug(
+          `Upgraded participant ${platformIdentity} → subscriber ${params.participantId} in conversation ${conversation._id}`
+        );
+      } else {
+        conversation.participants.push({ type: params.participantType, id: params.participantId });
+      }
+    } else {
+      conversation.participants.push({ type: params.participantType, id: params.participantId });
     }
+
+    await this.conversationRepository.updateParticipants(
+      params.environmentId,
+      params.organizationId,
+      conversation._id,
+      conversation.participants
+    );
   }
 }
